Carrito implementado

+++ Lab11_GTICS_Base_2023-2/src/main/java/edu/pucp/gtics/lab11_gtics_20232/entity/CarritoCompras.java
@@ -17,7 +17,7 @@ public class CarritoCompras {
     private int idcarritocompras;
 
     @Column(name = "cantidadtotal")
-    private int cantidadTotal;
+    private double cantidadTotal;
 
     @OneToOne
     @JoinColumn(name = "idusuario")
